On js

+++ src/Controller/TaskController.php
@@ -8,75 +8,91 @@ use App\Repository\TaskRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 class TaskController extends AbstractController
 {
     #[Route("/tasks", methods: 'POST')]
     public function store(
-        TaskRepository     $taskRepository,
+        TaskRepository $taskRepository,
         TaskListRepository $taskListRepository,
-        Request            $request
-    )
+        Request        $request
+    ): Response
     {
         $taskListId = $request->get('taskListId');
         $taskList = $taskListRepository->find($taskListId);
 
-        $this->denyAccessUnlessGranted('edit', $taskList->getBoard());
-
-        $boardId = $taskList->getBoard()->getId();
-
-        $taskId = (int)$request->get('taskId');
-        if ($taskId === 0) {
-            $task = new Task();
-        } else {
-            $task = $taskRepository->find($taskId);
-        }
+        $task = new Task();
 
         $task->setTaskList($taskList);
         $task->setText($request->get('text'));
         $taskRepository->add($task, true);
+        $taskId = $task->getId();
 
-        return $this->redirect("/boards/{$boardId}");
+        return $this->json([
+            'status' => 'ok',
+            'data' => [
+                'taskId' => $taskId,
+                'taskListId' => $taskListId
+            ]],
+            200, [], ['groups' => ['taskLists']]
+        );
     }
 
-    #[Route("/task-lists/{taskListId}/task/create", methods: 'GET')]
-    public function createTask(
-        TaskListRepository $taskListRepository,
-        int                $taskListId,
+    #[Route("/tasks/{id}", methods: 'GET')]
+    public function getTask(
+        TaskRepository $taskRepository,
+        Request         $request,
     )
     {
-        $taskList = $taskListRepository->find($taskListId);
+        $id = $request->get('id');
+        $task = $taskRepository->find($id);
 
-        $this->denyAccessUnlessGranted('edit', $taskList->getBoard());
+        $this->denyAccessUnlessGranted('view', $task->getTaskList()->getBoard());
 
-        return $this->render('todolist/task-edit.html.twig',
-            [
-                'taskId' => 0,
-                'taskList' => $taskList,
-                'taskText' => ""
-            ]);
+        return $this->json([
+            'status' => 'ok',
+            'data' => [
+                'task' => $task,
+                'taskListId' => $task->getTaskList()->getId(),
+                'boardId' => $task->getTaskList()->getBoard()->getId(),
+            ]],
+            200, [], ['groups' => ['taskList', 'taskLists']]
+        );
     }
 
-    #[Route("/tasks/{id}/edit", methods: 'GET')]
+    #[Route("/tasks/{id}", methods: 'PUT')]
     public function editTask(
-        TaskRepository $repository,
-        int            $id,
+        TaskRepository $taskRepository,
+        Task           $task,
+        Request        $request
     ): Response
     {
-        $task = $repository->find($id);
+        if ($request->get('text')) {
+            $task->setText($request->get('text'));
+        }
+
+        if ($request->get('isDone') !== null) {
+            $task->setIsDone($request->get('isDone'));
+        }
+
+        $taskRepository->add($task, true);
 
         $this->denyAccessUnlessGranted('edit', $task->getTaskList()->getBoard());
 
-        return $this->render('todolist/task-edit.html.twig',
-            [
-                'taskId' => $task->getId(),
-                'taskList' => $task->getTaskList(),
-                'taskText' => $task->getText()
-            ]);
+        return $this->json([
+            'status' => 'ok',
+            'data' => [
+                'taskListId' => $task->getTaskList()->getId(),
+                'taskText' => $task->getText(),
+                'isDone' => $task->isIsDone(),
+            ]],
+            200, [], ['groups' => ['taskLists']]
+        );
     }
 
-    #[Route("/tasks/{id}/remove")]
+    #[Route("/tasks/{id}", methods: 'DELETE')]
     public function removeTask(
         TaskRepository $taskRepository,
         int            $id
@@ -84,11 +100,14 @@ class TaskController extends AbstractController
     {
         $task = $taskRepository->find($id);
 
+        if ($task === null) {
+            throw new NotFoundHttpException();
+        }
+
         $this->denyAccessUnlessGranted('edit', $task->getTaskList()->getBoard());
 
-        $boardId = $task->getTaskList()->getBoard()->getId();
         $taskRepository->remove($task, true);
 
-        return $this->redirect("/boards/{$boardId}");
+        return $this->json(['status' => 'ok']);
     }
 }
